add associated constants to text component
