+++ ui/components/component-library/text/README.mdx
@@ -22,6 +22,8 @@ The `Text` accepts all props below as well as all [Box](/docs/ui-components-ui-b
 
 Use the `variant` prop and the `TEXT` object from `./ui/helpers/constants/design-system.js` to change the font size of the component.
 
+Optional: `TEXT` from ./text object can be used as instead.
+
 <Canvas>
   <Story id="ui-components-component-library-text-text-stories-js--variant" />
 </Canvas>
@@ -45,6 +47,8 @@ import { Text, TEXT } from '../../ui/components/component-library';
 ### Color
 
 Use the `color` prop and the `COLORS` object from `./ui/helpers/constants/design-system.js` to change the color of the `Text` component.
+
+Optional: `TEXT_COLORS` from ./text object can be used as instead.
 
 <Canvas>
   <Story id="ui-components-component-library-text-text-stories-js--color" />
@@ -103,6 +107,8 @@ Use the `fontWeight` prop and the `FONT_WEIGHT` object from `./ui/helpers/consta
 - `FONT_WEIGHT.MEDIUM` = `medium` || `500`
 - `FONT_WEIGHT.BOLD` = `bold` || `700`
 
+Optional: `FONT_WEIGHT` from ./text object can be used as instead.
+
 <Canvas>
   <Story id="ui-components-component-library-text-text-stories-js--font-weight" />
 </Canvas>
@@ -129,6 +135,8 @@ Use the `fontStyle` prop and the `FONT_STYLE` object from `./ui/helpers/constant
 - `FONT_STYLE.NORMAL`
 - `FONT_STYLE.ITALIC`
 
+Optional: `FONT_STYLE` from ./text object can be used as instead.
+
 <Canvas>
   <Story id="ui-components-component-library-text-text-stories-js--font-style" />
 </Canvas>
@@ -148,6 +156,8 @@ import { FONT_STYLE } from '../../../helpers/constants/design-system';
 ### Text Transform
 
 Use the `textTransform` prop and the `TEXT_TRANSFORM` object from `./ui/helpers/constants/design-system.js` to change the text alignment of the `Text` component
+
+Optional: `TEXT_TRANSFORM` from ./text object can be used as instead.
 
 <Canvas>
   <Story id="ui-components-component-library-text-text-stories-js--text-transform" />
@@ -171,6 +181,8 @@ import { TEXT_TRANSFORM } from '../../../helpers/constants/design-system';
 ### Text Align
 
 Use the `textAlign` prop and the `TEXT_ALIGN` object from `./ui/helpers/constants/design-system.js` to change the text alignment of the `Text` component
+
+Optional: `TEXT_ALIGN` from ./text object can be used as instead.
 
 <Canvas>
   <Story id="ui-components-component-library-text-text-stories-js--text-align" />
@@ -200,6 +212,8 @@ import { TEXT_ALIGN } from '../../../helpers/constants/design-system';
 ### Overflow Wrap
 
 Use the `overflowWrap` prop and the `OVERFLOW_WRAP` object from `./ui/helpers/constants/design-system.js` to change the overflow wrap of the `Text` component
+
+Optional: `OVERFLOW_WRAP` from ./text object can be used as instead.
 
 <Canvas>
   <Story id="ui-components-component-library-text-text-stories-js--overflow-wrap" />
