+++ ui/components/component-library/text/text.constants.js
@@ -1,13 +1,19 @@
-export const TEXT = {
-  DISPLAY_MD: 'display-md',
-  HEADING_LG: 'heading-lg',
-  HEADING_MD: 'heading-md',
-  HEADING_SM: 'heading-sm',
-  BODY_LG_MEDIUM: 'body-lg-medium',
-  BODY_MD: 'body-md',
-  BODY_MD_BOLD: 'body-md-bold',
-  BODY_SM: 'body-sm',
-  BODY_SM_BOLD: 'body-sm-bold',
-  BODY_XS: 'body-xs',
-  INHERIT: 'inherit',
+import {
+  FONT_STYLE,
+  FONT_WEIGHT,
+  OVERFLOW_WRAP,
+  TEXT,
+  TEXT_ALIGN,
+  TEXT_COLORS,
+  TEXT_TRANSFORM,
+} from '../../../helpers/constants/design-system';
+
+export {
+  FONT_STYLE,
+  FONT_WEIGHT,
+  OVERFLOW_WRAP,
+  TEXT,
+  TEXT_ALIGN,
+  TEXT_COLORS,
+  TEXT_TRANSFORM,
 };
