+++ ui/helpers/constants/design-system.js
@@ -152,6 +152,20 @@ export const TYPOGRAPHY = {
   Paragraph: 'p',
 };
 
+export const TEXT = {
+  DISPLAY_MD: 'display-md',
+  HEADING_LG: 'heading-lg',
+  HEADING_MD: 'heading-md',
+  HEADING_SM: 'heading-sm',
+  BODY_LG_MEDIUM: 'body-lg-medium',
+  BODY_MD: 'body-md',
+  BODY_MD_BOLD: 'body-md-bold',
+  BODY_SM: 'body-sm',
+  BODY_SM_BOLD: 'body-sm-bold',
+  BODY_XS: 'body-xs',
+  INHERIT: 'inherit',
+};
+
 const NONE = 'none';
 
 export const SIZES = {
